Keep smooth-lighting AO corners fractional instead of whole light levels
Smooth lighting averaged four neighbour samples in light-level space with
integer division, then rounded the AO-strength blend, so every corner landed
on a whole level. Because the b1.7.3 luminance curve drops 21% between level
15 and 14, {15,15,15,14}, {15,15,14,14} and {15,14,14,14} all collapsed onto
level 14: a face had nothing between fully lit and one hard step down, which
read as banding rather than ambient occlusion.

Carry the corner average as a float through BlockFaceRenderState and
Tessellator, and pack level*16 into vaUV2 with rounding at the last moment.
The attribute is an unnormalized ushort2 and the lightmap is GL_LINEAR, so a
fractional level lands between texels and the filter reproduces the vanilla
averaged-luminance value exactly for quarter-level averages.

The AO strength slider also fades continuously now rather than snapping a
whole level at a time.

+++ src/net/minecraft/client/render/Tessellator.hpp
@@ -46,7 +46,11 @@ class Tessellator {
  void color(int r, int g, int b, int a);
  void color(int rgb);
  void color(int rgb, int a);
- void light(int blockLight, int skyLight);
+ // Light levels are fractional: smooth lighting averages four neighbours, and
+ // rounding that average to a whole level collapses distinct AO corners onto
+ // the same lightmap texel. vaUV2 carries level*16, so quarter-level averages
+ // survive the pack exactly.
+ void light(float blockLight, float skyLight);
  void disableColor();
  void normal(float x, float y, float z);
  void blockData(double x,
@@ -102,8 +106,8 @@ class Tessellator {
  double blockCenterY_ = 0.0;
  double blockCenterZ_ = 0.0;
  int blockEmission_ = 0;
- int blockLight_ = 15;
- int skyLight_ = 15;
+ float blockLight_ = 15.0f;
+ float skyLight_ = 15.0f;
  int blockId_ = 0;
  bool blockFluid_ = false;
  int blockMetadata_ = 0;
